moving tests to the spec file
