+++ dpp/src/tests/document/state_transitions/validation/mod.rs
@@ -1,3 +1,4 @@
 mod validate_documents_batch_transition_state_spec;
+mod validate_documents_batch_transitions_basic_spec;
 mod validate_documents_uniqueness_by_indices_spec;
 mod validate_partial_compound_indices_spec;
